// MCP API/gen/cmp/types/v4/uuid_pb.ts
// @generated by protoc-gen-es v2.10.1 with parameter "target=ts,import_extension=.js"
// @generated from file cmp/types/v4/uuid.proto (package cmp.types.v4, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import { file_buf_validate_validate } from "../../../buf/validate/validate_pb.js";
import type { Timestamp } from "@bufbuild/protobuf/wkt";
import { file_google_protobuf_timestamp } from "@bufbuild/protobuf/wkt";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file cmp/types/v4/uuid.proto.
 */
export const file_cmp_types_v4_uuid: GenFile = /*@__PURE__*/
  fileDesc("ChdjbXAvdHlwZXMvdjQvdXVpZC5wcm90bxIMY21wLnR5cGVzLnY0Ih8KBFVVSUQSFwoFdmFsdWUYASABKAlCCLpIBXIDsAEBIm4KDEV4cGlyaW5nVVVJRBImCgJpZBgBIAEoCzISLmNtcC50eXBlcy52NC5VVUlEQga6SAPIAQESNgoKZXhwaXJhdGlvbhgCIAEoCzIaLmdvb2dsZS5wcm90b2J1Zi5UaW1lc3RhbXBCBrpIA8gBAWIGcHJvdG8z", [file_buf_validate_validate, file_google_protobuf_timestamp]);

/**
 * This must be a UUID according to RFC 4122 standard.
 *
 * ![Diagram](https://storage.googleapis.com/docs-cmp-files/diagrams/c4t/proto/cmp/types/v4/uuid.proto.dot.xs.svg)
 *
 * [Open Message Diagram](https://storage.googleapis.com/docs-cmp-files/diagrams/c4t/proto/cmp/types/v4/uuid.proto.dot.svg)
 *
 * @generated from message cmp.types.v4.UUID
 */
export type UUID = Message<"cmp.types.v4.UUID"> & {
  /**
   * This ensures that value is a valid [RFC 4122](https://datatracker.ietf.org/doc/html/rfc4122#section-4.1.2)
   *
   * @generated from field: string value = 1;
   */
  value: string;
};

/**
 * Describes the message cmp.types.v4.UUID.
 * Use `create(UUIDSchema)` to create a new message.
 */
export const UUIDSchema: GenMessage<UUID> = /*@__PURE__*/
  messageDesc(file_cmp_types_v4_uuid, 0);

/**
 * Generic timed UUID with an explicit expiration for follow up processes using this ID.
 *
 * @generated from message cmp.types.v4.ExpiringUUID
 */
export type ExpiringUUID = Message<"cmp.types.v4.ExpiringUUID"> & {
  /**
   * Unique ID
   *
   * @generated from field: cmp.types.v4.UUID id = 1;
   */
  id?: UUID;

  /**
   * Validity of the ID. When the timestamp is hit following requests using this
   * ID will fail as it is no longer valid. For example, a search ID
   * may only be valid for 15 minutes after which it cannot be used anymore for
   * validation requests.
   *
   * @generated from field: google.protobuf.Timestamp expiration = 2;
   */
  expiration?: Timestamp;
};

/**
 * Describes the message cmp.types.v4.ExpiringUUID.
 * Use `create(ExpiringUUIDSchema)` to create a new message.
 */
export const ExpiringUUIDSchema: GenMessage<ExpiringUUID> = /*@__PURE__*/
  messageDesc(file_cmp_types_v4_uuid, 1);